menu logic put in own file
removed the menu code from game into new menu.py file
